NavBar icon add, with msg
