+++ package.json
@@ -18,6 +18,7 @@
     "react": "^17.0.2",
     "react-bootstrap": "^2.0.0-rc.0",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
     "react-scripts": "4.0.3",
     "three": "^0.133.1",
     "typescript": "^4.4.3",
