+++ src/Components/Footer.tsx
@@ -1,0 +1,17 @@
+
+
+
+
+
+
+export default function Header (){
+    return(
+    <footer>
+     <h1 hidden>Mortum Producciones</h1>
+    <h3>Empresa enfocada a la realización de espectáculos de terror, obras de teatro, recorridos, contenido audiovisual e investigación paranormal. </h3>
+    <p>Y en la profunda oscuridad permanací largo tiempo atónito, temeroso...soñando sueños que ningún mortal se haya atrevido a soñar jamás --Edgar Allan Poe</p>
+     <br/>
+    
+ </footer>)
+ 
+ }
